Use DialogResult instead of EventsListener to progress code based on user's actions in the UI
In order to be able to start and close multiple revit transactions during the running of the application,
the transactions must all stay within the one single run of the revit plugin execute() method.
In order to achieve this, the execute() method must be able to pause waiting for the actions of the user.
This cannot be achieved by using external event handlers as the the execute() method will terminate before the event handlers will run.
Therefore, the only way to solve the problem is using the DialogResult() method on all the forms of the UI.
This forces the execute() method to pause until the user finishes using the form and, therefore, transactions can be run all along the run of the application.